move that show context data to individual modules
