+++ app/ptxboa_functions.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 """Utility functions for streamlit app."""
-from urllib.parse import urlparse
 
-import numpy as np
 import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
@@ -970,360 +968,6 @@ def create_infobox(context_data: dict):
     write_info(info2)
     write_info(info3)
     write_info(info4)
-
-
-def import_context_data():
-    """Import context data from excel file."""
-    filename = "data/context_data.xlsx"
-    cd = {}
-    cd["demand_countries"] = pd.read_excel(
-        filename, sheet_name="demand_countries", skiprows=1
-    )
-    cd["certification_schemes_countries"] = pd.read_excel(
-        filename, sheet_name="certification_schemes_countries"
-    )
-    cd["certification_schemes"] = pd.read_excel(
-        filename, sheet_name="certification_schemes", skiprows=1
-    )
-    cd["sustainability"] = pd.read_excel(filename, sheet_name="sustainability")
-    cd["supply"] = pd.read_excel(filename, sheet_name="supply", skiprows=1)
-    cd["literature"] = pd.read_excel(filename, sheet_name="literature")
-    cd["infobox"] = pd.read_excel(
-        filename,
-        sheet_name="infobox",
-        usecols="A:F",
-        skiprows=1,
-    ).set_index("country_name")
-    return cd
-
-
-def create_fact_sheet_demand_country(context_data: dict):
-    # select country:
-    country_name = st.session_state["country"]
-    with st.expander("What is this?"):
-        st.markdown(
-            """
-**Country fact sheets**
-
-This sheet provides you with additional information on the production and import of
- hydrogen and derivatives in all PTX BOA supply and demand countries.
-For each selected supply and demand country pair, you will find detailed
- country profiles.
-
- For demand countries, we cover the following aspects:
- country-specific projected hydrogen demand,
- target sectors for hydrogen use,
- hydrogen-relevant policies and competent authorities,
- certification and regulatory frameworks,
- and country-specific characteristics as defined in the demand countries'
- hydrogen strategies.
-
- For the supplying countries, we cover the country-specific technical potential
- for renewables (based on various data sources),
- LNG export and import infrastructure,
- CCS potentials,
- availability of an H2 strategy
- and wholesale electricity prices.
-            """
-        )
-    df = context_data["demand_countries"]
-    data = df.loc[df["country_name"] == country_name].iloc[0].to_dict()
-
-    flags_to_country_names = {
-        "France": ":flag-fr:",
-        "Germany": ":flag-de:",
-        "Netherlands": ":flag-nl:",
-        "Spain": ":flag-es:",
-        "China": ":flag-cn:",
-        "India": ":flag-in:",
-        "Japan": ":flag-jp:",
-        "South Korea": ":flag-kr:",
-        "USA": ":flag-us:",
-    }
-
-    st.subheader(
-        f"{flags_to_country_names[country_name]} Fact sheet for {country_name}"
-    )
-    with st.expander("**Demand**"):
-        c1, c2, c3 = st.columns(3)
-        with c1:
-            st.markdown("**Projected H2 demand in 2030:**")
-            st.markdown(data["h2_demand_2030"])
-            st.markdown(f"*Source: {data['source_h2_demand_2030']}*")
-        with c2:
-            st.markdown("**Targeted sectors (main):**")
-            st.markdown(data["demand_targeted_sectors_main"])
-            st.markdown(f"*Source: {data['source_targeted_sectors_main']}*")
-        with c3:
-            st.markdown("**Targeted sectors (secondary):**")
-            st.markdown(data["demand_targeted_sectors_secondary"])
-            st.markdown(f"*Source: {data['source_targeted_sectors_secondary']}*")
-
-    with st.expander("**Hydrogen strategy**"):
-        st.markdown("**Documents:**")
-        st.markdown(data["h2_strategy_documents"])
-
-        st.markdown("**Authorities:**")
-        st.markdown(data["h2_strategy_authorities"])
-
-    with st.expander("**Hydrogen trade characteristics**"):
-        st.markdown(data["h2_trade_characteristics"])
-        st.markdown(f"*Source: {data['source_h2_trade_characteristics']}*")
-
-    with st.expander("**Infrastructure**"):
-        st.markdown("**LNG import terminals:**")
-        st.markdown(data["lng_import_terminals"])
-        st.markdown(f"*Source: {data['source_lng_import_terminals']}*")
-
-        st.markdown("**H2 pipeline projects:**")
-        st.markdown(data["h2_pipeline_projects"])
-        st.markdown(f"*Source: {data['source_h2_pipeline_projects']}*")
-
-    if data["certification_info"] != "":
-        with st.expander("**Certification schemes**"):
-            st.markdown(data["certification_info"])
-            st.markdown(f"*Source: {data['source_certification_info']}*")
-
-
-def create_fact_sheet_supply_country(context_data: dict):
-    """Display information on a chosen supply country."""
-    # select country:
-    country_name = st.session_state["region"]
-    df = context_data["supply"]
-    data = df.loc[df["country_name"] == country_name].iloc[0].to_dict()
-
-    st.subheader(f"Fact sheet for {country_name}")
-    text = (
-        "**Technical potential for renewable electricity generation:**\n"
-        f"- {data['source_re_tech_pot_EWI']}: "
-        f"\t{data['re_tech_pot_EWI']:.0f} TWh/a\n"
-        f"- {data['source_re_tech_pot_PTXAtlas']}: "
-        f"\t{data['re_tech_pot_PTXAtlas']:.0f} TWh/a\n"
-    )
-
-    st.markdown(text)
-
-    text = (
-        "**LNG infrastructure:**\n"
-        f"- {data['lng_export']} export terminals\n"
-        f"- {data['lng_import']} import terminals.\n\n"
-        f"*Source: {data['source_lng']}*"
-    )
-
-    st.markdown(text)
-
-    st.write("TODO: CCS pot, elec prices, H2 strategy")
-
-
-def create_fact_sheet_certification_schemes(context_data: dict):
-    with st.expander("What is this?"):
-        st.markdown(
-            """
-**Get supplementary information on H2-relevant certification frameworks**
-
-This sheet provides you with an overview of current governmental regulations
-and voluntary standards for H2 products.
-            """
-        )
-    df = context_data["certification_schemes"]
-    helptext = "Select the certification scheme you want to know more about."
-    scheme_name = st.selectbox("Select scheme:", df["name"], help=helptext)
-    data = df.loc[df["name"] == scheme_name].iloc[0].to_dict()
-
-    # replace na with "not specified":
-    for key in data:
-        if data[key] is np.nan:
-            data[key] = "not specified"
-
-    st.markdown(data["description"])
-
-    with st.expander("**Characteristics**"):
-        st.markdown(
-            f"- **Relation to other standards:** {data['relation_to_other_standards']}"
-        )
-        st.markdown(f"- **Geographic scope:** {data['geographic_scope']}")
-        st.markdown(f"- **PTXBOA demand countries:** {data['ptxboa_demand_countries']}")
-        st.markdown(f"- **Labels:** {data['label']}")
-        st.markdown(f"- **Lifecycle scope:** {data['lifecycle_scope']}")
-
-        st.markdown(
-            """
-**Explanations:**
-
-- Info on "Geographical scope":
-  - This field provides an answer to the question: if you want to address a specific
- country of demand, which regulations and/or standards exist in this country
-   that require or allow proof of a specific product property?
-- Info on "Lifecycle scope":
-  - Well-to-gate: GHG emissions are calculated up to production.
-  - Well-to-wheel: GHG emissions are calculated up to the time of use.
-  - Further information on the life cycle scopes can be found in
-IRENA & RMI (2023): Creating a global hydrogen market: certification to enable trade,
- pp. 15-19
-"""
-        )
-
-    with st.expander("**Scope**"):
-        if data["scope_emissions"] != "not specified":
-            st.markdown("- **Emissions:**")
-            st.markdown(data["scope_emissions"])
-
-        if data["scope_electricity"] != "not specified":
-            st.markdown("- **Electricity:**")
-            st.markdown(data["scope_electricity"])
-
-        if data["scope_water"] != "not specified":
-            st.markdown("- **Water:**")
-            st.markdown(data["scope_water"])
-
-        if data["scope_biodiversity"] != "not specified":
-            st.markdown("- **Biodiversity:**")
-            st.markdown(data["scope_biodiversity"])
-
-        if data["scope_other"] != "not specified":
-            st.markdown("- **Other:**")
-            st.markdown(data["scope_other"])
-
-    with st.expander("**Sources**"):
-        st.markdown(data["sources"])
-
-
-def create_content_sustainability(context_data: dict):
-    with st.expander("What is this?"):
-        st.markdown(
-            """
-**Get supplementary information on PTX-relevant sustainability issues**
-
-Hydrogen is not sustainable by nature.
-And sustainability goes far beyond the CO2-footprint of a product.
-It also includes other environmental as well as socio-economic dimensions.
-
-This is why we provide you with a set of questions that will help you assess your plans
-for PTX production and export from a comprehensive sustainability perspective.
-Please note that this list does not claim to be exhaustive,
-but only serves for an orientation on the topic.
-            """
-        )
-    df = context_data["sustainability"]
-
-    c1, c2 = st.columns([2, 1])
-    with c1:
-        st.image("static/sustainability.png")
-        captiontext = (
-            "Source: https://ptx-hub.org/wp-content/uploads/2022/05/"
-            "PtX-Hub-PtX.Sustainability-Dimensions-and-Concerns-Scoping-Paper.pdf"
-        )
-        st.caption(captiontext)
-    with c2:
-        st.markdown(
-            """
-**Dimensions of sustainability**
-
-**What sustainability aspects should be considered for PTX products,
- production and policies?**
-
-**What questions should be asked before and during project development?**
-
-In this tab we aim to provide a basic approach to these questions.
- To the left, you can see the framework along which the compilation
- of sustainability aspects in this tab is structured. It is based on the EESG framework
- as elaborated by the PtX Hub and sustainability criteria developed by the Öko-Institut.
-
-**The framework distinguishes four key sustainability dimensions - Environmental,
- Economic, Social and Governance - from which you can select below.**
-
- Within each of these dimensions there are different clusters of sustainability aspects
- that we address in a set of questions. We differentiate between questions indicating
- guardrails and questions suggesting goals.
-
-With this compilation, we aim to provide a general overview of the sustainability
- issues that may be relevant in the context of PTX production. Of course,
- different aspects are more or less important depending on the project,
- product and country.
-
-**Take a look for yourself to see which dimensions are most important
- from where you are coming from.**
-                    """
-        )
-    st.divider()
-
-    c1, c2 = st.columns(2)
-    with c1:
-        helptext = "helptext"
-        dimension = st.selectbox(
-            "Select dimension:", df["dimension"].unique(), help=helptext
-        )
-    with c2:
-        helptext = """
-We understand **guardrails** as guidelines which can help you to produce green
-PTX products that are sustainable also beyond their greenhouse gas emission intensity.
-
-**Goals** are guidelines which can help link PTX production to improving local
- ecological and socio-economic circumstances in the supply country.
-They act as additional to guardrails which should be fulfilled in the first place
- to meet basic sustainability needs.
-"""
-        question_type = st.radio(
-            "Guardrails or goals?",
-            ["Guardrails", "Goals"],
-            help=helptext,
-            horizontal=True,
-        )
-        data = df.loc[(df["dimension"] == dimension) & (df["type"] == question_type)]
-
-    for topic in data["topic"].unique():
-        with st.expander(f"**{topic}**"):
-            data_select = data.loc[data["topic"] == topic]
-            for _ind, row in data_select.iterrows():
-                st.markdown(f"- {row['question']}")
-
-
-def is_valid_url(url: str) -> bool:
-    """Check if a string is a valid url."""
-    if not isinstance(url, str):
-        return False
-
-    try:
-        result = urlparse(url)
-        # Check if result.scheme and result.netloc are non-empty
-        return all([result.scheme, result.netloc])
-    except ValueError:
-        return False
-
-
-def create_content_literature(context_data: dict):
-    with st.expander("What is this?"):
-        st.markdown(
-            """
-**List of references**
-
-This tab contains a list of references used in this app.
-            """
-        )
-    df = context_data["literature"]
-    markdown_text = ""
-    for _ind, row in df.iterrows():
-        if is_valid_url(row["url"]):
-            text = f"- {row['long_name']}: [Link]({row['url']})\n"
-        else:
-            text = f"- {row['long_name']}\n"
-        markdown_text = markdown_text + text
-
-    st.markdown(markdown_text)
-
-
-def content_disclaimer():
-    with st.expander("What is this?"):
-        st.markdown(
-            """
-**Disclaimer**
-
-Information on product details of the PTX Business Opportunity Analyser
- including a citation suggestion of the tool.
-            """
-        )
-    st.image("static/disclaimer.png")
-    st.image("static/disclaimer_2.png")
 
 
 def config_number_columns(df: pd.DataFrame, **kwargs) -> {}:
